api/requirements.txt
Flask==1.0.3
flask_restful==0.3.7
gunicorn==19.9.0
mysqlclient==1.4.2.post1
flask_sqlalchemy==2.3.2
pyjwt==1.7.1
Flask-Mail==0.9.1
flask-socketio==4.0.0
eventlet==0.25.0
Flask-Caching==1.7.2
pycrypto==2.6.1
simple-crypt==4.1.7
apscheduler==3.6.0
